ajustes do bootstrap
