+++ GestaoReceita/.vs/SolutionWebCadastroLogin/config/applicationhost.config
@@ -163,7 +163,7 @@
             </site>
             <site name="SolutionWebCadastroLogin" id="3">
                 <application path="/" applicationPool="Clr4IntegratedAppPool">
-                    <virtualDirectory path="/" physicalPath="C:\Users\gscho\Documentos\Visual Studio 2022\Repos\Senac.GestaoReceita.Web\GestaoReceita\Web" />
+                    <virtualDirectory path="/" physicalPath="C:\Users\Aluno\Documents\Visual Studio 2022\repos\Senac.GestaoReceita.Web\GestaoReceita\Web" />
                 </application>
                 <bindings>
                     <binding protocol="http" bindingInformation="*:1105:localhost" />
