+++ GestaoReceita/Web/Controllers/HomeController.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using SolutionWebCadastroLogin.Controllers;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
@@ -6,7 +7,7 @@ using System.Web.Mvc;
 
 namespace Web.Controllers
 {
-    public class HomeController : Controller
+    public class HomeController : LoginExtention
     {
         public ActionResult Index()
         {
