+++ GestaoReceita/Web/Views/Login/Cadastro.cshtml
@@ -1,10 +1,8 @@
 ﻿@model SolutionWebCadastroLogin.Models.UsuariosCadastro.CadastroViewModel
 
 @{
-    Layout = "~/Views/Shared/_LayoutLogin.cshtml";
     ViewBag.Title = "Cadastro";
 }
-
 
 @using (Html.BeginForm("PersistirCadastro", "Login", FormMethod.Post))
 {
@@ -55,6 +53,28 @@
                 </form>
             </div>
         </div>
+        @if (!string.IsNullOrEmpty(this.Model.mensagemErro))
+        {
+            <div class="row">
+                <div class="col">
+                    <div class="alert alert-danger alert-dismissible fade show" role="alert">
+                        <strong>Erro!</strong> @this.Model.mensagemErro.
+                        <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+                    </div>
+                </div>
+            </div>
+        }
+        @if (!string.IsNullOrEmpty(this.Model.mensagemSucesso))
+        {
+            <div class="row">
+                <div class="col">
+                    <div class="alert alert-success alert-dismissible fade show" role="alert">
+                        <strong>Erro!</strong> @this.Model.mensagemSucesso.
+                        <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+                    </div>
+                </div>
+            </div>
+        }
     </div>
 
 }
